+++ pywincffi/kernel32/__init__.py
@@ -9,6 +9,8 @@ Provides functions, constants and utilities that wrap functions provided by
 # Our kernel32 package is broken into several submodules.  The functions
 # we're wrapping are imported here so it's easier to access and because
 # it's close to the way Windows would present them (as a single module)
+from pywincffi.kernel32.error import (
+    GetLastError, SetLastError, get_error_message)
 from pywincffi.kernel32.file import (
     ReadFile, WriteFile, FlushFileBuffers, MoveFileEx, CreateFile, LockFileEx,
     UnlockFileEx, GetTempPath)