+++ docs/source/changelog.rst
@@ -17,6 +17,11 @@ Notable enhancements and changes are:
       to provide improved support for overlapped IO. This is a **breaking change**
       because the return value and signature of the ReadFile function has
       changed.
+    * As part of :issue:`141` the following functions have been added:
+
+        * :func:`pywincffi.kernel32.error.GetLastError`
+        * :func:`pywincffi.kernel32.error.SetLastError`
+        * :func:`pywincffi.kernel32.error.get_error_message`
 
 0.5.0
 ~~~~~